separate xcuitest related methos in the xcuitest directory
