+++ lib/appium_lib/driver.rb
@@ -24,7 +24,11 @@ require_relative 'ios/element/generic'
 require_relative 'ios/element/textfield'
 require_relative 'ios/element/text'
 require_relative 'ios/mobile_methods'
-require_relative 'ios/xcuitest_gestures'
+
+# ios - xcuitest
+require_relative 'ios/xcuitest/element'
+require_relative 'ios/xcuitest/gestures'
+require_relative 'ios/xcuitest/mobile_methods'
 
 # android
 require_relative 'android/helper'
@@ -393,7 +397,8 @@ module Appium
         extend Appium::Android
       else
         extend Appium::Ios
-        if automation_name_is_xcuitest? # Override touch actions
+        if automation_name_is_xcuitest?
+          # Override touch actions and patch_webdriver_element
           extend Appium::Ios::Xcuitest
           extend Appium::Ios::Xcuitest::Gesture
         end
