+++ lib/appium_lib/ios/xcuitest/element.rb
@@ -1,0 +1,24 @@
+module Appium
+  module Ios
+    module Xcuitest
+      # @private
+      # class_eval inside a method because class Selenium::WebDriver::Element
+      # will trigger as soon as the file is required. in contrast a method
+      # will trigger only when invoked.
+      def patch_webdriver_element
+        Selenium::WebDriver::Element.class_eval do
+          # Enable access to iOS accessibility label
+          # accessibility identifier is supported as 'name'
+          def label
+            attribute('label')
+          end
+
+          # Cross platform way of entering text into a textfield
+          def type(text)
+            send_keys text
+          end # def type
+        end # Selenium::WebDriver::Element.class_eval
+      end # def patch_webdriver_element
+    end # module Xcuitest
+  end # module Ios
+end # module Appium
