+++ lib/appium_lib/ios/xcuitest/gestures.rb
